+++ load.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+cat $1 | nvd2sqlite3