+++ vfeed/lib/__init__.py
@@ -1,3 +1,0 @@
-import common
-import core
-import migration